requirements.txt
chromadb
sentence-transformers
docling
python-dotenv
python-telegram-bot
pydantic
supabase
